fix: Checking for link existence when deleting and updating

+++ internal/store/SQLstore/linkrepository.go
@@ -19,6 +19,7 @@ var (
 	InvalidLinkError = errors.New("The passed link is not valid")
 	ExistLinkError   = errors.New("The desired link already exists")
 	NoExistLinkError = errors.New("Such link does not exist")
+	WrongUserError   = errors.New("This link was created by another user")
 )
 
 // Заменить вызов двух функций на транзакцию
@@ -171,6 +172,22 @@ func (l *LinkRepository) ShortLinkExist(ctx context.Context, shortLink string) (
 	return false, nil
 }
 
+func (l *LinkRepository) OriginLinkExist(ctx context.Context, originLink string) (bool, error) {
+	ctx, cancel := context.WithTimeout(ctx, l.store.QueryTimeout)
+	defer cancel()
+	var count int
+	if err := l.store.db.QueryRowContext(ctx,
+		"SELECT COUNT(*) FROM links WHERE OriginLink = $1;",
+		originLink,
+	).Scan(&count); err != nil {
+		return true, err
+	}
+	if count > 0 {
+		return true, nil
+	}
+	return false, nil
+}
+
 func (l *LinkRepository) LinkExistByLinkAndUser(ctx context.Context, originLink string, uid int) (bool, error) {
 	ctx, cancel := context.WithTimeout(ctx, l.store.QueryTimeout)
 	defer cancel()
@@ -191,6 +208,13 @@ func (l *LinkRepository) LinkExistByLinkAndUser(ctx context.Context, originLink 
 func (l *LinkRepository) Delete(ctx context.Context, originLink string) error {
 	ctx, cancel := context.WithTimeout(ctx, l.store.QueryTimeout)
 	defer cancel()
+	exist, err := l.OriginLinkExist(ctx, originLink)
+	if err != nil {
+		return err
+	}
+	if !exist {
+		return NoExistLinkError
+	}
 	if err := l.store.db.QueryRowContext(ctx,
 		"DELETE FROM links WHERE OriginLink = $1;",
 		originLink,
@@ -203,12 +227,19 @@ func (l *LinkRepository) Delete(ctx context.Context, originLink string) error {
 func (l *LinkRepository) DeleteByUser(ctx context.Context, originLink string, uid int) error {
 	ctx, cancel := context.WithTimeout(ctx, l.store.QueryTimeout)
 	defer cancel()
-	exist, err := l.LinkExistByLinkAndUser(ctx, originLink, uid)
+	exist, err := l.OriginLinkExist(ctx, originLink)
 	if err != nil {
 		return err
 	}
 	if !exist {
 		return NoExistLinkError
+	}
+	exist, err = l.LinkExistByLinkAndUser(ctx, originLink, uid)
+	if err != nil {
+		return err
+	}
+	if !exist {
+		return WrongUserError
 	}
 	if err := l.store.db.QueryRowContext(ctx,
 		"DELETE FROM links WHERE OriginLink = $1 AND UID = $2;",
@@ -223,12 +254,19 @@ func (l *LinkRepository) DeleteByUser(ctx context.Context, originLink string, ui
 func (l *LinkRepository) ReActivate(ctx context.Context, originLink string, uid int) error {
 	ctx, cancel := context.WithTimeout(ctx, l.store.QueryTimeout)
 	defer cancel()
-	exist, err := l.LinkExistByLinkAndUser(ctx, originLink, uid)
+	exist, err := l.OriginLinkExist(ctx, originLink)
 	if err != nil {
 		return err
 	}
 	if !exist {
 		return NoExistLinkError
+	}
+	exist, err = l.LinkExistByLinkAndUser(ctx, originLink, uid)
+	if err != nil {
+		return err
+	}
+	if !exist {
+		return WrongUserError
 	}
 	if err := l.store.db.QueryRowContext(ctx,
 		"UPDATE links SET Status = $1 WHERE OriginLink = $2 AND UID = $3;",
